pyproject.toml
[project]
name = "openclaw-doctor"
version = "0.1.0"
description = "A CLI tool to diagnose, validate, and auto-fix OpenClaw AI assistant installations"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "OpenClaw Community"}
]
keywords = ["openclaw", "cli", "doctor", "diagnostics", "ai-assistant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
openclaw-doctor = "openclaw_doctor.main:app"

[project.urls]
Homepage = "https://github.com/openclaw-community/openclaw-doctor"
Repository = "https://github.com/openclaw-community/openclaw-doctor"
Issues = "https://github.com/openclaw-community/openclaw-doctor/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openclaw_doctor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
